requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-limiter[redis]==3.7.0
PyJWT==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
redis==5.0.4
